Renderer->GameContainer, move unit scaling to level scene.  Add basic scene management
