+++ src/game-container.js
@@ -1,0 +1,100 @@
+// 1.6:1 aspect ratio
+export const ASPECT_RATIO = 1.6/1;
+
+export default class GameContainer {
+    constructor(canvas, initialScene) {
+        this.canvas = canvas;
+        this.ctx = canvas.getContext('2d');
+        this.stopping = false;
+
+        this.gameLoop = this.gameLoop.bind(this);
+        this.resizeCanvas = this.resizeCanvas.bind(this);
+        this.lastFrameTime = Date.now();
+        this.FPS = 0;
+
+        this.sceneStack = [];
+
+        if (initialScene) {
+            this.pushScene(initialScene);
+        }
+    }
+
+    pushScene(scene) {
+        this.sceneStack.push(scene);
+        scene.container = this;
+    }
+
+    resizeCanvas() {
+        this.width = this.canvas.clientWidth;
+        this.height = this.canvas.clientHeight;
+
+        this.canvas.width = this.width;
+        this.canvas.height = this.height;
+
+        if (this.canvas.height) {
+            const ratio = this.width / this.height;
+
+            if (ratio > ASPECT_RATIO) {
+                // Too wide
+                this.width = ASPECT_RATIO * this.height;
+            } else {
+                // Too thin
+                this.height = this.width / ASPECT_RATIO;
+            }
+        }
+    }
+
+    preDraw() {
+        this.ctx.setTransform(1, 0, 0, 1, 0, 0);
+
+        // This probably won't be needed once we start drawing
+        //  a proper background
+        this.ctx.clearRect(0, 0, this.canvas.width, this.canvas.height);
+
+        this.ctx.translate((this.canvas.width - this.width) / 2, (this.canvas.height - this.height) / 2);
+
+        this.ctx.save();
+
+    }
+
+    postDraw() {
+        this.ctx.restore();
+
+        const now = Date.now();
+        this.ctx.fillStyle = '#FF0000';
+        this.ctx.textAlign = 'right';
+        this.ctx.font = '16px sans-serif';
+
+        this.FPS = this.FPS * 0.9 + (now - this.lastFrameTime) * 0.1;
+        this.ctx.fillText(`FPS: ${Math.round(1000 / this.FPS)}`, this.width - 10, 16);
+
+        this.lastFrameTime = now;
+    }
+
+    gameLoop() {
+        this.preDraw();
+
+        for (const scene of this.sceneStack) {
+            scene.draw(this.ctx);
+        }
+
+        this.postDraw();
+
+        if (!this.stopping) {
+            window.requestAnimationFrame(this.gameLoop);
+        } else {
+            this.stopping = false;
+        }
+    }
+
+    start() {
+        window.requestAnimationFrame(this.gameLoop);
+        window.addEventListener('resize', this.resizeCanvas);
+        this.resizeCanvas();
+    }
+
+    stop() {
+        this.stopping = true;
+        window.removeEventListener('resize', this.resizeCanvas);
+    }
+}
